requirements.txt
Django==4.2.16
PyMySQL==1.0.2
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
aiosqlite==0.21.0
httpx==0.28.1
langchain==1.2.13
langgraph==1.1.3
pycryptodome==3.23.0
